refactor remote access tests, mark strict=false
